Merge branch ci (#15)
Combine ci and cd actions into a single action to make the process faster. The cd part of the action is conditional on the ci part.

Build a docker image on the runner instead of pulling an image from the docker registry, which is slow and fragile.

Modify the entrypoint for the container to tail /dev/null to keep the container running indefinitely.

Delete emscripten and macos action metadata files. The final ci/cd scheme uses a linux container running gcc to create the linux build for testing and the emscripten build for deployment

Change the library path for linux build to be appropriate for an amd64 machine. This is a cosmetic change to remove a warning message from the container startup logs about mismatch between the host and container architecture

Update docker compose to remove the the explicit tail command since tailing /dev/null is now the default entry point for the container.